Revert "[CHANGE] Created material resource pool class"
This reverts commit 3cea347ebf5231134388453c580a322a89d04c84.
