+++ src/d3d12/d3d12_renderer.cpp
@@ -7,8 +7,7 @@
 #include "../window.hpp"
 
 #include "d3d12_defines.hpp"
-#include "d3d12_resource_pool_material.hpp"
-#include "d3d12_resource_pool_model.hpp"
+#include "d3d12_resource_pool.hpp"
 #include "d3d12_functions.hpp"
 #include "d3d12_pipeline_registry.hpp"
 #include "d3d12_shader_registry.hpp"
