+++ src/resource_pool.cpp
@@ -2,6 +2,32 @@
 
 namespace wr
 {
-	
+
+	MaterialPool::MaterialPool(std::size_t size_in_mb) : m_size_in_mb(size_in_mb)
+	{
+	}
+
+	//! Loads a material
+	MaterialHandle MaterialPool::Load(std::string_view path, TextureType type)
+	{
+		return MaterialHandle();
+	}
+
+	ModelPool::ModelPool(std::size_t size_in_mb) : m_size_in_mb(size_in_mb)
+	{
+	}
+
+	//! Loads a model without materials
+	Model* ModelPool::Load(std::string_view path, ModelType type)
+	{
+		return new Model();
+	}
+
+	//! Loads a model with materials
+	std::pair<Model*, std::vector<MaterialHandle>> ModelPool::LoadWithMaterials(MaterialPool* material_pool, std::string_view path, ModelType type)
+	{
+		auto model = new Model();
+		return std::pair<Model*, std::vector<MaterialHandle>>();
+	}
 
 } /* wr */
