+++ src/scene_graph/scene_graph.hpp
@@ -6,8 +6,7 @@
 #include <DirectXMath.h>
 
 #include "../util/defines.hpp"
-#include "../resource_pool_material.hpp"
-#include "../resource_pool_model.hpp"
+#include "../resource_pool.hpp"
 
 namespace wr
 {
